Remove IClassType interface
