--- PHP/Types/Models/IClassType.php
@@ -1,7 +1,0 @@
-<?php
-namespace PHP\Types\Models;
-
-/**
- * Defines an interface for class types
- */
-interface IClassType extends IType {}
